Page add group

+++ src/Controller/AdminController.php
@@ -7,10 +7,12 @@ use App\Entity\Group;
 use App\Entity\Media;
 use App\Entity\Trick;
 use App\Entity\User;
+use App\Form\GroupFormType;
 use App\Form\MediaFormType;
 use App\Form\RegistrationFormType;
 use App\Form\TrickFormType;
 use App\Repository\CommentRepository;
+use App\Repository\GroupRepository;
 use App\Repository\TrickRepository;
 use App\Repository\UserRepository;
 use DateTime;
@@ -35,13 +37,11 @@ class AdminController extends AbstractController
      * @IsGranted("ROLE_ADMIN", message="Vous n'avez pas l'autorisation d'accèder à cette page")
      * @Route("/admin", name="app_admin")
      */
-    public function index(TrickRepository $trickRepository): Response
+    public function index(Request $request, TrickRepository $trickRepository): Response
     {
-        if (!in_array('ROLE_ADMIN', $this->getUser()->getRoles(), false)) $this->redirect('home');
-
         $nbTricks = $trickRepository->countAllTricks();
         $nbPages = $trickRepository->getNbPagesTricks($nbTricks);
-        $page = $_GET['page'] ?? 1;
+        $page = $request->query->get('page') ?? 1;
         $tricks = $trickRepository->getTricksPages($page, $nbPages);
 
         return $this->render('admin/index.html.twig', [
@@ -55,11 +55,11 @@ class AdminController extends AbstractController
      * @IsGranted("ROLE_ADMIN", message="Vous n'avez pas l'autorisation d'accèder à cette page")
      * @Route("/admin/users", name="admin_users")
      */
-    public function usersAdmin(UserRepository $userRepository): Response
+    public function usersAdmin(Request $request, UserRepository $userRepository): Response
     {
         $nbUsers = $userRepository->countAllUsers();
         $nbPages = $userRepository->getNbPagesUsers($nbUsers);
-        $page = $_GET['page'] ?? 1;
+        $page = $request->query->get('page') ?? 1;
         $users = $userRepository->getUsersPages($page, $nbPages);
 
         return $this->render('admin/admin_users.html.twig', [
@@ -76,17 +76,16 @@ class AdminController extends AbstractController
     public function userComments(int $idUser, Request $request, EntityManagerInterface $entityManager, CommentRepository $commentRepository): Response
     {
         $user = $this->entityManager->getRepository(User::class)->find($idUser);
-        $nbComments = $commentRepository->countCommentByUser($user);
+        $nbComments = $commentRepository->countCommentsByUser($user);
         $nbPages = $commentRepository->getNbPagesComments($nbComments);
-        $page = $_GET['page'] ?? 1;
+        $page = $request->query->get('page') ?? 1;
         $comments = $commentRepository->getCommentsByUserPages($page, $nbPages, $idUser);
-
-//        $comments = $this->entityManager->getRepository(Comment::class)->findBy( ['user_id' => $idUser]);
 
         return $this->render('admin/admin_user_comments.html.twig', [
             'user' => $user,
             'nbPages' => $nbPages,
-            'currentPage' => $page
+            'currentPage' => $page,
+            'comments' => $comments
        ]);
     }
 
@@ -97,16 +96,16 @@ class AdminController extends AbstractController
     public function tricksComments(int $idTrick, Request $request, EntityManagerInterface $entityManager, CommentRepository $commentRepository): Response
     {
         $trick = $this->entityManager->getRepository(Trick::class)->find($idTrick);
-        $nbComments = $commentRepository->countCommentByTrick($trick);
+        $nbComments = $commentRepository->countCommentsByTrick($trick);
         $nbPages = $commentRepository->getNbPagesComments($nbComments);
-        $page = $_GET['page'] ?? 1;
+        $page = $request->query->get('page') ?? 1;
         $comments = $commentRepository->getCommentsByTrickPages($page, $nbPages, $idTrick);
-
 
         return $this->render('admin/admin_trick_comments.html.twig', [
             'trick' => $trick,
             'nbPages' => $nbPages,
-            'currentPage' => $page
+            'currentPage' => $page,
+            'comments' => $comments
         ]);
     }
 
@@ -114,8 +113,12 @@ class AdminController extends AbstractController
      *  @IsGranted("ROLE_ADMIN", message="Vous n'avez pas l'autorisation d'accèder à cette page")
      * @Route("/admin/comment/{idComment}/change", name="admin_comments_change")
      */
-    public function changeCommentStatus(int $idComment, Request $request, EntityManagerInterface $entityManager)
+    public function changeCommentStatus(Request $request, int $idComment, EntityManagerInterface $entityManager)
     {
+        $page = '';
+        if ($request->get('page') !== null) {
+            $page = '?page=' . $request->get('page');
+        }
         $comment = $this->entityManager->getRepository(Comment::class)->find($idComment);
 
         if ($comment->getIsVerified() === true) {
@@ -129,13 +132,13 @@ class AdminController extends AbstractController
 
         $this->addFlash(
             'success',
-            'Le statut du commentaire à été modifié avec succès !'
+            'La vérification du commentaire à été modifié avec succès !'
         );
 
-        if (isset($_POST['trick'])) {
-            $url = '/admin/trick/' . $_POST['trick'] . '/comments';
-        } elseif (isset($_POST['user'])) {
-            $url = '/admin/users/' . $_POST['user'] . '/comments';
+        if ($request->get('trick') !== null) {
+            $url = '/admin/trick/' . $request->get('trick') . '/comments' . $page;
+        } elseif ($request->get('user') !== null) {
+            $url = '/admin/users/' . $request->get('user') . '/comments' . $page;
         }
 
         return $this->redirect($url);
@@ -147,6 +150,10 @@ class AdminController extends AbstractController
      */
     public function changeUserStatus(int $idUser, Request $request, EntityManagerInterface $entityManager)
     {
+        $page = '';
+        if ($request->get('page') !== null) {
+            $page = '?page=' . $request->get('page');
+        }
         $user = $this->entityManager->getRepository(User::class)->find($idUser);
 
         if (in_array('ROLE_ADMIN', $user->getRoles())) {
@@ -160,10 +167,10 @@ class AdminController extends AbstractController
 
         $this->addFlash(
             'success',
-            "Le statut de l'utilisateur " . $user->getUsername() . " à été modifié avec succès !"
+            "Le rôle de l'utilisateur " . $user->getUsername() . " à été modifié avec succès !"
         );
 
-        return $this->redirectToRoute('admin_users');
+        return $this->redirect('/admin/users' . $page);
     }
 
     /**
@@ -173,7 +180,6 @@ class AdminController extends AbstractController
     public function deleteTrick(int $idTrick, Request $request, EntityManagerInterface $entityManager)
     {
         $trick = $this->entityManager->getRepository(Trick::class)->findOneBy(['id' => $idTrick]);
-
 
         //Suppression de tout les media du trick
         foreach ($trick->getMedia() as $media)
@@ -194,7 +200,7 @@ class AdminController extends AbstractController
     }
 
     /**
-     * @Route("/trick/new", name="add_trick")
+     * @Route("/trick/add", name="add_trick")
      * @IsGranted("ROLE_USER")
      */
     public function addTrick(Request $request, EntityManagerInterface $entityManager)
@@ -295,23 +301,36 @@ class AdminController extends AbstractController
      */
     public function deleteMedia(int $idMedia, EntityManagerInterface $entityManager)
     {
-        //AJOUTER TRY CATCH
         $media = $this->entityManager->getRepository(Media::class)->find($idMedia);
         $trick = $media->getTrick();
+        $typeMedia = $media->getType();
+        //On s'assure qu'il y ait bien au moins une image et une vidéo
+        $mediasSameType = 0;
+        foreach ($trick->getMedia() as $media) {
+            if ($media->getType() == $typeMedia) {
+                $mediasSameType++;
+            }
+            if($mediasSameType > 1 ) {
+                break;
+            }
+        }
 
-        dump($media);
-        dump($trick);
+        if ($mediasSameType==1) {
+            $this->addFlash(
+                'warning',
+                "Erreur : Vous ne pouvez pas supprimer ce media car il s'agit de la seule " . $media->getType()
+            );
+        } else {
+            $trick->removeMedium($media);
+            $entityManager->remove($media);
 
-        $trick->removeMedium($media);
-        $entityManager->remove($media);
+            $entityManager->flush();
 
-        $entityManager->flush();
-
-        $this->addFlash(
-            'success',
-            'La média à été supprimé avec succès !'
-        );
-
+            $this->addFlash(
+                'success',
+                'La média à été supprimé avec succès !'
+            );
+        }
         return $this->redirect('/trick/' . $trick->getId() . '-' . $trick->getSlug() . '/edit');
     }
 
@@ -376,4 +395,116 @@ class AdminController extends AbstractController
             'trick' => $trick,
         ]);
     }
+
+    /**
+     * @Route("/group/add", name="add_group")
+     * @IsGranted("ROLE_USER")
+     */
+    public function addGroup(Request $request, EntityManagerInterface $entityManager)
+    {
+        $group = new Group();
+
+        if ($request->query->get('url') !== null) {
+            $url = "/trick/" . $request->query->get('url') . "/edit";
+        } else {
+            $url = '/trick/add';
+        }
+
+        $form = $this->createForm(GroupFormType::class, $group);
+
+        $form->handleRequest($request);
+
+        if ($form->isSubmitted() && $form->isValid()) {
+            $group = $form->getData();
+
+            $entityManager->persist($group);
+            $entityManager->flush();
+            $this->addFlash(
+                'success',
+                'La groupe à été ajouté avec succès !'
+            );
+            return $this->redirect($url);
+        }
+
+        return $this->render('admin/add_group.html.twig', [
+            'form' => $form->createView(),
+            'group' => $group,
+        ]);
+    }
+
+    /**
+     * @Route("/admin/groups/", name="admin_group")
+     * @IsGranted("ROLE_ADMIN")
+     */
+    public function displayGroup(Request $request, GroupRepository $groupRepository): Response
+    {
+        $nbTricks = $groupRepository->countAllGroups();
+        $nbPages = $groupRepository->getNbPagesGroups($nbTricks);
+        $page = $request->query->get('page') ?? 1;
+        $groups = $groupRepository->getGroupsPages($page, $nbPages);
+
+        return $this->render('admin/group.html.twig', [
+            'groups' => $groups,
+            'nbPages' => $nbPages,
+            'currentPage' => $page
+        ]);
+    }
+
+    /**
+     * @Route("/admin/group/{idGroup}/edit", name="edit_group")
+     * @IsGranted("ROLE_ADMIN")
+     */
+    public function editGroup(int $idGroup,Request $request, EntityManagerInterface $entityManager)
+    {
+        $group = $this->entityManager->getRepository(Group::class)->find($idGroup);
+
+        $form = $this->createForm(GroupFormType::class, $group);
+
+        $form->handleRequest($request);
+
+        if ($form->isSubmitted() && $form->isValid()) {
+            $group = $form->getData();
+
+            $entityManager->persist($group);
+            $entityManager->flush();
+            $this->addFlash(
+                'success',
+                'La groupe "' . $group->getTitle() . '" à été modifié avec succès !'
+            );
+            return $this->redirectToRoute('admin_group');
+        }
+
+        return $this->render('admin/edit_group.html.twig', [
+            'form' => $form->createView(),
+            'group' => $group,
+        ]);
+    }
+
+    /**
+     * @Route("/admin/group/{idGroup}/delete", name="delete_group")
+     * @IsGranted("ROLE_ADMIN")
+     */
+    public function deleteGroup(int $idGroup, Request $request, EntityManagerInterface $entityManager)
+    {
+        $group = $this->entityManager->getRepository(Group::class)->find($idGroup);
+
+        if ($group->getTricks()->isEmpty()) {
+            $entityManager->remove($group);
+            $entityManager->flush();
+
+            $this->addFlash(
+                'success',
+                'Le groupe ' . $group->getTitle() . ' à été supprimé avec succès !'
+            );
+        } else {
+
+            $this->addFlash(
+                'danger',
+                'Le groupe "' . $group->getTitle() . '" est associé à des figures et ne peut pas être supprimé.'
+            );
+        }
+        return $this->redirectToRoute('admin_group');
+    }
+
+
 }
